requirements.txt
temporalio
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.24.0